// CaseStudy_CentralLibraryTracker/CentralLibraryTrackerFrontEnd/src/app/models/book-issue.ts
export class BookIssue {
    id: number;
    bookId: number;
    memberId: number;
    subject: string;
    title: string;
    author: string;
    publisher: string;
    copyright: string;
    edition: number;
    pages: number;
    isbn: number;
    libraryName: string;
    name: string;
    email: string;
    major: string;
    issueDate: Date;
    expireDate: Date;
}